+++ pygls/protocol.py
@@ -361,9 +361,9 @@ class JsonRPCProtocol(asyncio.Protocol):
         if error is not None:
             logger.debug('Received error response to message "%s": %s', msg_id, error)
             future.set_exception(JsonRpcException.from_dict(error))
-
-        logger.debug('Received result for message "%s": %s', msg_id, result)
-        future.set_result(result)
+        else:
+            logger.debug('Received result for message "%s": %s', msg_id, result)
+            future.set_result(result)
 
     def _procedure_handler(self, message):
         """Delegates message to handlers depending on message type."""